Admin page refactoring
