+++ components/Admin/TournamentRings.tsx
